% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ggplot.R
\name{as.ggplot}
\alias{as.ggplot}
\title{as.ggplot}
\usage{
as.ggplot(plot)
}
\arguments{
\item{plot}{base or grid plot, or graphic generated by ggplot, lattice, etc.}
}
\value{
ggplot object
}
\description{
convert plot to ggplot object
}
\author{
guangchuang yu
}
