% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-grob.R
\name{as.grob}
\alias{as.grob}
\title{as.grob}
\usage{
as.grob(plot)
}
\arguments{
\item{plot}{base or grid plot, or graphic object generated by ggplot, lattice, etc.}
}
\value{
grob object
}
\description{
convert a plot to grob object
}
\author{
guangchuang yu
}
