% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{autocomplete_name}
\alias{autocomplete_name}
\title{autocomplete_name}
\usage{
autocomplete_name(name, ...)
}
\arguments{
\item{name}{partial text name}

\item{...}{additional parameters}
}
\value{
scientific name
}
\description{
obtaion suggestions for full names based on partial text name
}
