% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{dev.off}
\alias{dev.off}
\title{gridSVG Graphics Device}
\usage{
dev.off(...)
}
\arguments{
\item{...}{
    Further parameters that are passed onto a \code{NULL}
    \code{\link{pdf}} graphics device. Useful parameters include
    \code{width} and \code{height}.
  }
}
\value{
\code{gridsvg} returns nothing.

  \code{dev.off} will return in the same manner as
  \code{\link[gridSVG]{grid.export}}. A list is always returned, but invisibly
  when an invalid filename is given.
}
\description{
Provides a convenient and familiar graphics device interface for the
  gridSVG package.
}
\details{
These functions provide a more familiar and perhaps convenient
  interface to gridSVG than \code{\link[gridSVG]{grid.export}}. It uses a PDF
  device as drawing occurs, but when the device needs to be written out
  (via \code{dev.off}) then it will save an SVG image instead.

  When a grid display list is not in use, or any device other than the
  \code{gridsvg} device is used, the behaviour of \code{dev.off} is the
  same as \code{\link[grDevices]{dev.off}} from the \code{grDevices}
  package.
}
\seealso{
\code{\link{pdf}} and \code{\link[gridSVG]{grid.export}}.
}
\author{
Simon Potter
}
