% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bgimage.R
\name{geom_bgimage}
\alias{geom_bgimage}
\title{geom_bgimage}
\usage{
geom_bgimage(image)
}
\arguments{
\item{image}{image file}
}
\value{
ggplot
}
\description{
add image as background to plot panel.
}
\author{
Guangchuang Yu
}
