% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nba.R
\name{geom_nba}
\alias{geom_nba}
\title{geom_nba}
\usage{
geom_nba(mapping = NULL, data = NULL, inherit.aes = TRUE, na.rm = FALSE,
  by = "width", ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{inherit.aes}{whether inherit aes mapping from ggplot()}

\item{na.rm}{whether remove NA values}

\item{by}{one of 'width' or 'height' for specifying size}

\item{...}{additional parameter}
}
\value{
ggplot2 layer
}
\description{
geom layer for using nba logo image
}
