% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{geom_phylopic}
\alias{geom_phylopic}
\title{geom_phylopic}
\usage{
geom_phylopic(mapping = NULL, data = NULL, inherit.aes = TRUE,
  na.rm = FALSE, by = "width", height = 512, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{inherit.aes}{whether inherit aes mapping from ggplot()}

\item{na.rm}{whether remove NA values}

\item{by}{one of 'width' or 'height' for specifying size}

\item{height}{size (by height) of phylopic image to be used}

\item{...}{additional parameter}
}
\value{
ggplot2 layer
}
\description{
geom layer for using phylopic image
}
\author{
guangchuang yu
}
