% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_subview.R
\name{geom_subview}
\alias{geom_subview}
\title{geom_subview}
\usage{
geom_subview(subview, x, y, width = 0.1, height = 0.1)
}
\arguments{
\item{subview}{subview}

\item{x}{x position}

\item{y}{y position}

\item{width}{width}

\item{height}{height}
}
\value{
layer
}
\description{
subview geom
}
\author{
guangchuang yu
}
