% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_twitchemote.R
\name{geom_twitchemote}
\alias{geom_twitchemote}
\title{geom_twitchemotes}
\usage{
geom_twitchemote(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{inherit.aes}{whether inherit aes mapping from ggplot()}

\item{na.rm}{whether remove NA values}

\item{by}{one of 'width' or 'height' for specifying size}

\item{...}{additional parameter}
}
\value{
ggplot2 layer
}
\description{
geom layer for using Twitch emotes
}
\author{
Brendan Rocks
}
