% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbackground.R
\name{ggbackground}
\alias{ggbackground}
\title{ggbackground}
\usage{
ggbackground(gg, background, ...)
}
\arguments{
\item{gg}{gg object}

\item{background}{background image}

\item{...}{additional parameter to manipulate background image, see also geom_image}
}
\value{
gg object
}
\description{
set background for ggplot
}
\author{
guangchuang yu
}
