% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggimage-package.R
\docType{package}
\name{ggimage-package}
\alias{ggimage}
\alias{ggimage-package}
\title{ggimage: Use Image in 'ggplot2'}
\description{
Supports image files and graphic objects to be visualized in 'ggplot2' graphic system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/ggimage}
  \item Report bugs at \url{https://github.com/YuLab-SMU/ggimage/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Yonghe Xia \email{xiayh17@gmail.com} [contributor]
}

}
\keyword{internal}
