% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{ggpreview}
\usage{
ggpreview(plot = last_plot(), width, height, units = "in")
}
\arguments{
\item{plot}{any plot that supported by the 'ggplotify' package}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{units of the 'width' and 'height'}
}
\value{
a preview of the figure
}
\description{
preview a plot befor saving it to a file.
}
\author{
Guangchuang Yu
}
