% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_read2.R
\name{image_read2}
\alias{image_read2}
\title{image_read2}
\usage{
image_read2(path, ..., cut_empty_space = TRUE)
}
\arguments{
\item{path}{file path}

\item{...}{additional parameters that pass to magick::image_read}

\item{cut_empty_space}{whether remove marginal empty space}
}
\value{
magick-image object
}
\description{
read image (by magick::image_read) with the ability to remove marginal empty space
}
\author{
Guangchuang Yu
}
