% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{phylopic_uid}
\alias{phylopic_uid}
\title{phylopic_uid}
\usage{
phylopic_uid(name, seed = 123)
}
\arguments{
\item{name}{scientific name}

\item{seed}{The random seed to use to generate the same uid,
because a name might have many uid, the function will extract one
of them randomly, default is 123.}
}
\value{
phylopic uid
}
\description{
query phylopic to get uid from scientific name
}
\author{
Guangchuang Yu
}
