% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_nothing}
\alias{theme_nothing}
\title{theme_nothing}
\usage{
theme_nothing(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{
ggplot2 theme
}
\description{
A theme that only show the plot panel
}
\author{
Guangchuang Yu
}
