% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{download_phylopic}
\alias{download_phylopic}
\title{download_phylopic}
\usage{
download_phylopic(id, destdir = ".", ...)
}
\arguments{
\item{id}{phylopic id}

\item{destdir}{directory where the downloaded images are to be saved.}

\item{...}{additional parameters passed to download.file}
}
\value{
a character string (or vector) with downloaded file path
}
\description{
download phylopic images
}
\details{
This function allows users to download phylopic images using phylopic id
}
\author{
Guangchuang Yu
}
