% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{download_phylopic}
\alias{download_phylopic}
\title{download_phylopic}
\usage{
download_phylopic(
  id,
  destdir = ".",
  ...,
  reuse = TRUE,
  lock_timeout = getOption("ggimage.phylopic_lock_timeout", 30),
  lock_poll = getOption("ggimage.phylopic_lock_poll", 0.05)
)
}
\arguments{
\item{id}{phylopic id}

\item{destdir}{directory where the downloaded images are to be saved.}

\item{...}{additional parameters passed to download.file}

\item{reuse}{whether to reuse existing non-empty regular files.}

\item{lock_timeout}{maximum seconds to wait for a concurrent download lock.}

\item{lock_poll}{seconds between lock acquisition attempts.}
}
\value{
a character string (or vector) with downloaded file path
}
\description{
download phylopic images
}
\details{
This function allows users to download phylopic images using phylopic id
}
\author{
Guangchuang Yu
}
