% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_image.R
\name{geom_image}
\alias{geom_image}
\title{geom_image}
\usage{
geom_image(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  is_cairo_svg = FALSE,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge image}

\item{is_cairo_svg}{logical, whether to process the image as cairo svg. See notes.}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer for visualizing image files
}
\note{
If image is compatible with the
\href{https://cran.r-project.org/package=grImport2}{{grImport2}} package, then
setting \code{is_cairo_svg = TRUE} and printing to \code{gridsvg()} can generate a
fully vectorized plot. See examples.
}
\examples{
\dontrun{
library("ggplot2")
library("ggimage")

set.seed(2017-02-21)

## raster
d <- data.frame(x = rnorm(10),
                y = rnorm(10),
                image = sample(c("https://www.r-project.org/logo/Rlogo.png",
                                "https://jeroenooms.github.io/images/frink.png"),
                              size=10, replace = TRUE)
               )
ggplot(d, aes(x, y)) + geom_image(aes(image=image))

## full vector
data <- data.frame(
  x = rnorm(10),
  y = rnorm(10),
  image = system.file("extdata/Rlogo_cairo.svg", package = "ggimage")
)

gridsvg("gg_full_vector.svg", strict = FALSE)
  data |>
    ggplot(aes(x, y)) +
    geom_image(aes(image = image), size = .1, is_cairo_svg = TRUE)
dev.off()
}
}
\author{
Guangchuang Yu
}
