% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{geom_phylopic}
\alias{geom_phylopic}
\title{geom_phylopic}
\usage{
geom_phylopic(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  image_fun = function(x) magick::image_background(x, color = "none", flatten = FALSE),
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{inherit.aes}{whether inherit aes mapping from ggplot()}

\item{na.rm}{whether remove NA values}

\item{by}{one of 'width' or 'height' for specifying size}

\item{image_fun}{a function to process magick-image objects, default is
\code{function(x)magick::image_background(x, color = "none", flatten = FALSE)}.}

\item{...}{additional parameter}
}
\value{
ggplot2 layer
}
\description{
geom layer for using phylopic image
}
\author{
Guangchuang Yu
}
