% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{ggpreview}
\usage{
ggpreview(
  filename = NULL,
  plot = last_plot(),
  width = NA,
  height = NA,
  units = "in",
  ...
)
}
\arguments{
\item{filename}{If it is not NULL, the previewed figure will be save to the file}

\item{plot}{any plot that supported by the 'ggplotify' package}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{units of the 'width' and 'height'}

\item{...}{additional parameters pass to ggsave() if filename is not NULL}
}
\value{
a preview of the figure
}
\description{
preview a plot befor saving it to a file.
}
\author{
Guangchuang Yu
}
