% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylopic.R
\name{phylopic_uid}
\alias{phylopic_uid}
\title{phylopic_uid}
\usage{
phylopic_uid(name)
}
\arguments{
\item{name}{scientific name}
}
\value{
phylopic uid
}
\description{
query phylopic to get uid from scientific name
}
\author{
Guangchuang Yu
}
