% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.grob}
\alias{reexports}
\alias{as.ggplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplotify}{\code{\link[ggplotify]{as.grob}}, \code{\link[ggplotify]{as.ggplot}}}
}}

